functions/requirements.txt
firebase-admin~=6.2.0
firebase-functions~=0.4.1
functions-framework~=3.4.0
python-shogi~=1.1.1
gymnasium==0.29.1
torch==2.3.1
numpy~=2.0.0
torch~=2.3.1